requirements.txt
appnope==0.1.4
asttokens==2.4.1
blinker==1.8.0
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
fonttools==4.51.0
imageio==2.34.1
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyzmq==26.0.2
scikit-image==0.23.2
scipy==1.13.0
six==1.16.0
stack-data==0.6.3
tifffile==2024.4.24
tornado==6.4
traitlets==5.14.3
wcwidth==0.2.13
Werkzeug==3.0.2
